requirements.txt
pymupdf
numpy==2.2.6
pdf2image==1.17.0
Pillow==11.2.1
streamlit==1.45.1
ultralytics==8.3.141
torch
onnxruntime
onnx
transformers
scikit-image